[ConstraintSystem] NFC: Move `ConstraintGraph{Scope}.h` to `include/swift/Sema`
